Adds OpenCL support for faster tree evaluation
Implements OpenCL-based tree evaluation to significantly improve performance.

Introduces a new OpenCLTreeRunner class that compiles and runs the evolved trees on an OpenCL device, leveraging the parallel processing capabilities of GPUs for faster computation.  This change includes modifications to the Node class and its implementations to generate OpenCL-compatible code. The scikit-learn compatible classifier and regressor are modified to use the OpenCL tree runner.

The change involves skipping the "Clover" platform in OpenCL as it does not support all features.

Also provides an updated example notebook for classification and regression.